Basic split of code base
