+++ src/affix_sub_graph.rs
@@ -1,0 +1,8 @@
+use handlegraph::handle::Handle;
+
+// structure for storing reported subgraph
+pub struct AffixSubgraph {
+    pub sequence: String,
+    pub parents: Vec<Handle>,
+    pub shared_prefix_nodes: Vec<Handle>,
+}
